backup/data_ingestion/requirements.txt
aiohttp
pandas 
datetime
asyncio
argparse
json 
airflow.providers.google.cloud.hooks.gcs
from google.cloud import storage